+++ docs/examples/multiframe.html
@@ -1,0 +1,78 @@
+<!DOCTYPE html>
+<html class="h-100 overflow-hidden">
+  <head>
+    <meta charset="UTF-8" />
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl"
+      crossorigin="anonymous"
+    />
+    <title>Larvitar - Multiframe rendering example</title>
+  </head>
+
+  <body class="h-100" style="background-color: #d7e8e1">
+    <div class="row h-100">
+      <div
+        id="viewer"
+        class="col-8 h-100"
+        style="background-color: black"
+      ></div>
+
+      <div class="col-4 h-100">
+        <pre class="h-100">
+          <code class="javascript" data-lang="javascript">
+          <p style="font-size:0.6vw;">
+            
+          </p>
+          </code>
+        </pre>
+      </div>
+    </div>
+
+    <script src="../../dist/larvitar.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.21/lodash.min.js"></script>
+
+    <script>
+
+      let demoFiles = [];
+   
+      // init all
+      larvitar.initLarvitarStore();
+      larvitar.initializeImageLoader();
+      larvitar.registerMultiFrameImageLoader();
+      larvitar.initializeCSTools();
+      larvitar.larvitar_store.addViewport("viewer");
+
+     
+
+      function renderSerie() {
+        larvitar.resetImageParsing();
+        larvitar.readFiles(demoFiles, function (seriesStack, err) {
+          // render the first series of the study
+          let seriesId = _.keys(seriesStack)[0];
+          let serie = seriesStack[seriesId];
+          console.log(serie);
+          larvitar.buildMultiFrameImage(seriesId, serie);
+          let multiFrameSerie = larvitar.multiFrameManager[seriesId];
+          larvitar.renderFrame(multiFrameSerie, "viewer")
+          // larvitar.renderImage(multiFrameSerie, "viewer");
+          // larvitar.renderImage(serie, "viewer");
+          larvitar.addDefaultTools();
+          larvitar.setToolActive(
+            larvitar.larvitar_store.state.leftMouseHandler
+          );
+        });
+      }
+
+      async function createFile(fileName, cb) {
+        let response = await fetch("./demo/" + fileName);
+        let data = await response.blob();
+        let file = new File([data], fileName);
+        demoFiles.push(file);
+        cb();
+      }
+      createFile("multiframe.dcm", renderSerie);
+    </script>
+  </body>
+</html>
